chore: move x509 to common package
