+++ apiml-security-common/src/main/java/org/zowe/apiml/security/common/util/X509Util.java
@@ -8,7 +8,7 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.gateway.x509;
+package org.zowe.apiml.security.common.util;
 
 import lombok.Value;
 import lombok.experimental.UtilityClass;
